+++ src/com/tutorial/main/QuickBoi.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public class QuickBoi extends GameObject{
     private Handler handler;
 
-    public QuickBoi(int x, int y, ID id, Handler handler) {
+    public QuickBoi(float x, float y, ID id, Handler handler) {
         super(x, y, id);
 
         velX = 2;
@@ -14,7 +14,7 @@ public class QuickBoi extends GameObject{
         this.handler = handler;
     }
     public Rectangle getBounds(){
-        return new Rectangle(x, y, 24, 24);
+        return new Rectangle((int)x,(int) y, 24, 24);
     }
     public void tick() {
         x += velX;
@@ -31,7 +31,7 @@ public class QuickBoi extends GameObject{
 
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillOval(x, y, 24, 24);
+        g.fillOval((int)x,(int) y, 24, 24);
     }
 }
